issue-tracker-cli: src/lib.rs module split (SA R13 F1 Trigger B closure)
Closes the final Solution Architect carry-forward finding (SA R13 F1
Trigger B): src/lib.rs grew from ~660 LOC pre-Layer-6 to ~1130 LOC
non-test (~2250 LOC including tests) and crossed the cognitive-load
threshold SA flagged across four review rounds. The "storage / validate
/ commands triad" split named in SA R3 F2 (Layer 1 dismissal),
re-affirmed in SA R6 F4 / SA R11 F1 / SA R13 F1 / SA R15 F1, and
auto-Backlogged per CLOSURE-PROTOCOL.md §3 at Layer 7 R2 — landed now.

New module structure:

- `src/storage.rs` (~200 lines) — data types + persistence + load-time
  invariants. Contains: `Tracker`, `Issue`, `load_tracker`, `save_tracker`,
  `tracker_is_valid`, `issue_fields_are_valid`, `description_is_valid`,
  `label_is_valid`, `parse_timestamp`, `CORRUPT_DATA_ERROR`,
  `VALID_STATUSES`, `PRIORITY_ORDER`. Load-time invariant predicates
  live here (rather than in validate.rs) so `load_tracker` can call
  them without crossing a module boundary; user-input validators live
  in validate.rs as a separate concern.
- `src/validate.rs` (~250 lines) — user-input validators + safety
  transforms + arithmetic/time helpers. Contains: `validate_title`,
  `validate_description`, `parse_status`, `parse_priority`,
  `parse_label`, `parse_id`, `dedupe_labels`, `bump_next_id`,
  `current_timestamp`, `display_safe`, `sanitize_quoted_values`.
  Imports `VALID_STATUSES` + `PRIORITY_ORDER` from storage as the
  single source of truth for the domain enums.
- `src/commands.rs` (~660 lines) — command implementations + rendering
  layer + color helpers. Contains: `CreateArgs`, `cmd_create`,
  `cmd_list`, `cmd_status`, `cmd_show`, `cmd_delete`, `ColorMode`,
  `color_mode_from_env`, all color/rendering helpers (`priority_ansi`,
  `status_ansi`, `wrap_color`, `render_cell`, `format_show_block`,
  `show_label`, `format_list_header`, `format_list_row`,
  `filter_issues`, `issue_matches_filters`, `truncate_with_ellipsis`,
  `sort_issues`, `priority_rank`, `label_matches`), and all
  column-width constants (`ID_WIDTH`, `STATUS_WIDTH`, `PRIORITY_WIDTH`,
  `LABELS_WIDTH`, `TITLE_WIDTH`, `LABEL_COLUMN_WIDTH`, `ANSI_RESET`).
- `src/lib.rs` (~50 lines hub + ~1080 lines tests) — module
  declarations + `pub use` re-exports preserving the pre-split public
  API surface. The `#[cfg(test)] mod tests` block remains intact (93
  unit tests cohabit at the lib.rs level) with `use crate::commands::*;`
  + `use crate::storage::*;` glob imports bringing `pub(crate)`
  internal helpers into the test scope.

Visibility:
- Cross-module-used internal helpers are `pub(crate)` so the test mod
  in lib.rs and other submodules can reach them without making them
  part of the public API.
- Public surface (`Tracker`, `Issue`, `CreateArgs`, `ColorMode`,
  `cmd_*`, `color_mode_from_env`, `validate_*`, `parse_*`,
  `load_tracker`, `save_tracker`, `display_safe`, `sanitize_quoted_values`,
  `bump_next_id`, `current_timestamp`, `dedupe_labels`, `label_matches`,
  `sort_issues`) is preserved via lib.rs `pub use` re-exports —
  `src/main.rs` and integration tests (which use unqualified paths
  like `tracker::cmd_create`, `tracker::CreateArgs`) need no changes.

No behavior change. The split is pure code reorganization:
- cargo test 237/237 pass (93 unit + 32+18+9+25+7+33+20 layer 1-7
  integration). No tests moved; no test bodies changed.
- cargo clippy --all-targets --locked -- -D warnings clean.
- cargo fmt --check clean.
- cargo audit 0 advisories.

Carry-forward state after this commit:

- VDD-IAR R17 F1 (CRITICAL Red Gate compliance) — Resolved (fbbb8a3).
- SA R11 F1 (cmd_list rendering extraction) — Resolved (3fa1f3c).
- SA R13 F1 Trigger B (src/lib.rs module split) — Resolved (this
  commit). Closes the entire SA carry-forward cluster that was
  auto-Backlogged §3 at Layer 7 R2.
- SA R13 F2 (format_show_block column constants) — Resolved (3fa1f3c).
- QE R17 F1 (force_color test seam) — Resolved (bd7511e).
- QE R17 F5 (CJK display-width latent risk) — Resolved (c341a54).
- Platform R12 F3 (clippy pre-commit hook) — Resolved (ff0e85c).

Every Round-1 deferred finding has now landed terminal closure.
